[page] finish commoditySpecificList page and it's api.

+++ src/components/commoditySpecificList.vue
@@ -24,7 +24,7 @@
     <div class="operation">
       <div class="operation-btns">
           <el-button type="primary" @click="productdAdd">新增库存</el-button>
-          <!-- <el-button type="danger"  @click="standardDeleteAll">删除</el-button> -->
+          <!-- <el-button type="danger"  @click="productDeleteAll">删除</el-button> -->
       </div>
     </div>
     <!-- /功能 -->
@@ -36,7 +36,7 @@
         <el-table-column prop="title" label="商品名称"></el-table-column>
         <el-table-column prop="attrs" label="具体规格" show-overflow-tooltip>
           <template slot-scope="scope">
-            <span v-for="(tag, index) in scope.row.attrs" :key="index" style="margin-right: 5px;">{{tag}}</span>
+            <span v-for="(tag, index) in scope.row.attrs" :key="tag.id" style="margin-right: 5px;">{{tag.title}}</span>
           </template>
         </el-table-column>
         <el-table-column prop="price" label="价格">
@@ -52,19 +52,13 @@
         <el-table-column prop="created_at" label="创建时间"></el-table-column>
         <el-table-column label="操作" width="150">
           <template slot-scope="scope">
-            <el-button size="small" type="primary" @click="commodityEdit(scope.$index, scope.row)">修改</el-button>
-            <el-button size="small" type="danger"  @click="standardDelete(scope.$index, scope.row)">删除</el-button>
+            <el-button size="small" type="primary" @click="productEdit(scope.row)">修改</el-button>
+            <el-button size="small" type="danger"  @click="productDelete(scope.$index, scope.row)">删除</el-button>
           </template>
         </el-table-column>
       </el-table>
     </div>
 
-    <!-- 页码 -->
-    <div class="pages">
-      <el-pagination @current-change="handleCurrentChange" :current-page.sync="currentPage" :page-size="10" layout="total, prev, pager, next" :total="count">
-      </el-pagination>
-    </div>
-    <!-- /页码 -->
   </section>
 </template>
 
@@ -83,7 +77,7 @@ export default {
       },
 
       currentPage: 1,
-      count: 0,
+      count: 20,
 
       //模拟数据
       commodities: []
@@ -127,13 +121,11 @@ export default {
      * @param {number} index 当前行（数据）索引
      * @param {object} row 当前行（数据）所有信息
      */
-    commodityEdit(index, row) {
-      console.log(index);
-    },
-
-    //添加库存
-    addCommodity(id) {
-      console.log(id);
+    productEdit(row) {
+      this.$router.push({
+        name: "commoditySpecificEdit",
+        params: { product: row }
+      });
     },
 
     /**
@@ -141,23 +133,27 @@ export default {
      * @param {number} index 当前行（数据）索引
      * @param {object} row 当前行（数据）所有信息
      */
-    standardDelete(index, row) {
-      console.log(index);
-      this.$api.deletecommodities(row.id, res => {
-        this.commodities.splice(index, 1);
-        this.$message({
-          type: "success",
-          message: "删除成功"
+    productDelete(index, row) {
+      this.$confirm("此操作将删除该商品库存, 是否继续?", "提示", {
+        confirmButtonText: "确定",
+        cancelButtonText: "取消",
+        type: "warning"
+      })
+        .then(() => {
+          this.$api.deletecommodities(row.id, res => {
+            this.commodities.splice(index, 1);
+            this.$message({
+              type: "success",
+              message: "删除成功"
+            });
+          });
+        })
+        .catch(() => {
+          this.$message({
+            type: "info",
+            message: "已取消删除"
+          });
         });
-      });
-    },
-
-    /**
-     * 页码事件
-     * @param {number} page 当前页码
-     * */
-    handleCurrentChange(page) {
-      console.log(page);
     }
   }
 };
